feat: split globals to support smaller worker chuncks
