+++ tns-core-modules/globals/polyfills/dialogs/package.json
@@ -1,0 +1,5 @@
+{
+  "name" : "dialogs",
+  "main" : "dialogs",
+  "nativescript": {}
+}
