+++ tns-core-modules/globals/polyfills/fetch/fetch.ts
@@ -1,0 +1,8 @@
+import "../../core";
+import "../../polyfills/xhr";
+
+import { installPolyfills } from "../polyfill-helpers";
+
+global.registerModule("fetch", () => require("../../../fetch"));
+
+installPolyfills("fetch", ["fetch", "Headers", "Request", "Response"]);
