+++ tns-core-modules/globals/polyfills/fetch/package.json
@@ -1,0 +1,5 @@
+{
+  "name" : "fetch",
+  "main" : "fetch",
+  "nativescript": {}
+}
