+++ tns-core-modules/globals/polyfills/polyfill-helpers/package.json
@@ -1,0 +1,5 @@
+{
+  "name" : "polyfill-helpers",
+  "main" : "polyfill-helpers",
+  "nativescript": {}
+}
