+++ tns-core-modules/globals/polyfills/timers/package.json
@@ -1,0 +1,5 @@
+{
+  "name" : "timers",
+  "main" : "timers",
+  "nativescript": {}
+}
