+++ tns-core-modules/globals/polyfills/timers/timers.ts
@@ -1,0 +1,6 @@
+import "../../core";
+import { installPolyfills } from "../polyfill-helpers";
+
+global.registerModule("timer", () => require("../../../timer"));
+
+installPolyfills("timer", ["setTimeout", "clearTimeout", "setInterval", "clearInterval"]);
