+++ tns-core-modules/globals/polyfills/xhr/package.json
@@ -1,0 +1,5 @@
+{
+  "name" : "xhr",
+  "main" : "xhr",
+  "nativescript": {}
+}
