+++ tns-core-modules/globals/polyfills/xhr/xhr.ts
@@ -1,0 +1,6 @@
+import "../../core";
+import { installPolyfills } from "../polyfill-helpers";
+
+global.registerModule("xhr", () => require("../../../xhr"));
+
+installPolyfills("xhr", ["XMLHttpRequest", "FormData"]);
